+++ src/map/enmity_container.cpp
@@ -153,7 +153,7 @@ void CEnmityContainer::AddPartyEnmity(CCharEntity* PChar)
 			}
 		}else if (PChar->PParty->m_PAlliance != NULL)
 				{
-					for (int32 a = 0; a < PChar->PParty->m_PAlliance->partyList.size(); ++a)
+					for (uint8 a = 0; a < PChar->PParty->m_PAlliance->partyList.size(); ++a)
 					{
 							for (uint8 i = 0; i < PChar->PParty->m_PAlliance->partyList.at(a)->members.size(); ++i)
 							{
